requirements.txt
streamlit>=1.28.0
anthropic>=0.24.0
pyyaml>=6.0
python-docx>=0.8.11
PyPDF2>=4.0.0
beautifulsoup4>=4.12.0
openpyxl>=3.11.0
pandas>=2.0.0
lxml>=4.9.0
